add LSSupportsOpeningDocumentsInPlace to Info.plist (fixes Xcode 10.1 compiler warning)
the LSSupportsOpeningDocumentsInPlace key is necessary because Little Go
declares in its Info.plist that it works as editor for files which have
the document type com.red-bean.sgf (which have file extension .sgf).
an application declaring file types must specify whether it can open
them in place or needs copies to work on.

apparently the Info.plist key LSSupportsOpeningDocumentsInPlace has
existed for a few years now (Apple added the key to the Info.plist
reference in September 2015), but Xcode has only now started to nag if
the key is missing.