code signing overhaul for Xcode 10.1
despite the carefully researched and documented (in NOTES.Build) code
signing approach chosen in June 2016 (see commit 9d0ea4b), code signing
once again was broken after a major Xcode upgrade. the Xcode 8.3 release
notes state that the user interface for managing signing certificates
and provisioning profiles changed, but I suspect that it wasn't only
the UI that changed, but other stuff as well.

because I couldn't get code signing to work manually, I decided to give
Xcode another chance to automatically manage code signing. this is a
setting that can be enabled on the "General" tab of the Xcode project
editor when the project "Little Go" is selected.

I also manually deleted the following build settings:
- On the project level
  PROVISIONING_PROFILE = "";
  "PROVISIONING_PROFILE[sdk=*]" = "<uuid>";
- On the Unit Test target level
  "PROVISIONING_PROFILE[sdk=*]" = "<uuid>";

the current settings work for deploying to / debugging on a device.
whether they also work for signing an AppStore build remains to be seen.