suicide check in stone setup no longer depends on order in which properties are set (#323)
the old suicide check only examined the liberties of the stones being
set up. this does not work in the case where the stones being set up
take liberties away from ***OTHER*** stones that are already there.

for instance, set up A1 with a black stone (property blackSetupStones
is being set), then set up A2 and B1 with white stones (property
whiteSetupStones is being set). The white stones put the black stone on
A1 into a suicidal position. this was not detected by the previous check
because it only looked at the liberties of A2 and B1.

the new suicide check looks at all stones on the board. the order in
which setup properties are set is therefore no longer important.