adapt to height change of UITableViewCellStyleSubtitle in iOS 11 (#314)
since iOS 11 cells with UITableViewCellStyleSubtitle have a different
height than cells of other types. in addition, the height is not the
same across all devices.

this affects BoardPositionTableListViewController (which is effective
on iPads) because that controller must calculate the total height of
a UITableView.

the new way of calculating the UITableView height can no longer be based
solely on the rendering of an offscreen cell, because offscreen cells
with style UITableViewCellStyleSubtitle do not have an increased height!
iOS apparently increases the height only when the cell is actually
rendered on sreen. the correct method of calculating the UITableView
height therefore requires a combination of 1) making some assumptions
about the height change in UiElementMetrics; and 2) implementing the
UITableViewDelegate method tableView:heightForRowAtIndexPath:(), the
latter in order to force UITableView to use the same cell height that
was used in calculating the UITableView height.