make unarchiving a GoGame reusable
this change was initiated because it was thought that the new command
UnarchiveGameCommand could be used from UserDefaultsUpdater. it turned
out that this is not the case, the timing during app launch does not
allow to create a new GoGame at the time that UserDefaultsUpdater
performs its task.

so although UnarchiveGameCommand is no longer needed at the moment, it
seemed a waste to throw the class away.