increase nscodingVersion to make sure that an NSCoding archive is discarded
in the upcoming commits a new application state is introduced: the mode
that the UI area "Play" is in. currently the app only distinguishes
between scoring mode and non-scoring mode (i.e. play mode). the new
mode allows more than just these two selections.

the UI area "Play" mode is also a user preference.

when the user launches the app the next time after the new mode is
introduced, and the NSCoding archive contains a state where scoring mode
is enabled, then the UI area "Play" mode in the user preferences should
be set correspondingly. normally it would be the job of
UserDefaultsUpdater to do this. it turns out, however, that
UserDefaultsUpdater cannot unarchive the GoGame to look at the scoring
mode, because of a timing issue during app launch.

so to make sure that scoring mode is guaranteed to be disabled when the
app is launched the next time, we have to force the app to discard the
NSCoding archive.