fix displaying "more game actions" menu as popover on iPad
the hack for finding the UIView associated with an UIBarButtonItem no
longer works on iOS 14 (and probably iOs 13, too) because navigation
bars no longer have the views that correspond to their UIBarButtonItems
as direct subviews. as a consequence the popover was displayed from the
wrong UIView/CGRect.

the solution is exquisitely simpler: UIPopoverPresentationController has
a barButtonItem property that can be assigned with the UIBarButtonItem
object that triggered the message, and UIPopoverPresentationController
automagically figures out where to present the popover from. this API
exists since iOS 8, apparently I was just too focused on the
UIView/CGRect solution to notice it.

because of this new elegant solution, there is no longer a need to query
the GameActionManagerUIDelegate for a view.