app no longer crashes when SGF file is imported (fixes #362)
issue found during testing on an iPad with iOS 14 installed.

the crash occurs because HandleDocumentInteractionCommand performs a UI
change in a background thread.  UIKit explicitly raises an exception
which then causes the app to crash:

  Modifications to the layout engine must not be performed from a
  background thread after it has been accessed from the main thread.

the exception is raised if document interaction occurs after the app has
been terminated. the exception is not raised if document interaction
occurs when the app merely has been suspended.

the exception is also not raised with the previous version of the app.
apparently building the app with the newest iOS SDK causes the exception
to be raised.

the exception not being raised does not automatically mean that the old
app version did not have a problem. actually there WERE some crash
reports related to document interaction, I just couldn't pinpoint the
root cause. this commit might therefore be a fix for those crashes.