Remove UIViewControllerBasedStatusBarAppearance from Info.plist
the property was introduced in commit
d65ac80c6ffa6db20be95c68c400725ca5a97980. the property was set to false
so that UIStatusBarStyle - which was also set in the same commit - can
take effect. in a later commit UIStatusBarStyle was removed, but
UIViewControllerBasedStatusBarAppearance remained in Info.plist.

by removing the property we revert back to the original behaviour before
it was introduced. according to this article
  https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/iPhoneOSKeys.html#//apple_ref/doc/uid/TP40009252-SW29
the property has default value YES.