detect board view smaller dimension instead of hardcoding it (#339)
hardcoding the smaller dimension depending on UIType and interface
orientation is no longer viable. with the new annotation view we are
more fully using the available space so that the varying screen sizes
of supported devices can lead to the board view having a different
smaller dimension for the same UIType / interface orientation (e.g.
iPad Pro vs. regular iPad in landscape orientation).