markupOnPointsDidChange now has correct data when markup is erased in a rectangle
the erroneous code did not interpret the result of
handleEraseMarkupOnIntersection:withNodeMarkup:(). if any of the points
in the rectangle contained markup, pointsWithChangedMarkup would get
set. on the next iteration, singleOrNoMarkupWasErased would get set to
false even if the point of the current iteration did not have any
markup.

the net effect of the erroneous code was that it was almost impossible
to have the markupOnPointsDidChange notification carry data for only
a single point with changed markup. it *was* possible, but only if the
point was the last in the iteration.

the changed and correct code now only sets pointsWithChangedMarkup
and/or singleOrNoMarkupWasErased if
handleEraseMarkupOnIntersection:withNodeMarkup:() actually deleted
something.