fix button box / annotation view support for Dark mode (#379)
setting the "alpha" property of a view causes the subviews to also be
rendered transparent. the resulting appearance is of the subviews being
"dimmed".

before commit e2b1b054c58bdec565124ddfccc0db3f5ce099a6 was made, the
original solution instead applied the transparency only to the
background color. subviews in this case are rendered crisp and clear.
this resulting appearance is vastly better, so this commit returns to
the original solution.

unfortunately, because this no longer uses a system-provided color that
automatically switches when dark mode is enabled/disabled, the involved
view controllers must now become active themselves and react to dark
mode being enabled/disabled.