UI layout unification: merge root VCs for UIAreaPlay for UILayoutPhone and UILayoutPad (#371)
as a result of the UI layout unification, both the .h and .m files of
PlayRootViewControllerPhone and PlayRootViewControllerPad were 99%
identical, line for line, so merging the two classes into one is
inevitable, although version control history of the original files is
lost by this action.

the 3 differences that existed are now encoded in the
switch-statement in the initializer of the new class
PlayRootViewControllerPhoneAndPad.