fix bug in move number drawing (introduced by erroneous optimization)
the optimization added earlier (and that this commit removes) did not
take into account that move numbers already drawn must not be overdrawn
when the intersection is encountered again in an earlier move.

this commit adds a new, albeit very small, optimization that avoids a
lookup in the pointsWithMarkup array not only if the intersection is
currently empty, but also if there is a stone on it from the other color
than the stone played by the move to be numbered.