fix view transition when game is loaded while game info screen is visible (#364)
the main issue fixed in this commit is that the game info screen was
popped in a secondary thread.

also, the delay when popping ViewGameController from the navigation
stack is no longer necessary due to the changes made during UI layout
unification.

finally, the sequence of operations in ViewGameController was adjusted
to avoid a debug warning
  "Unbalanced calls to begin/end appearance transitions"
in iOS 13 and newer. for iOS 12 the debug warning is still emitted for
unknown reasons.