add initial implementation of ResizableStackViewController (#380)
the upcoming variation view will be another scroll view that competes
with the board view for screen estate. clearly the user must be able to
resize the views to his liking. ResizableStackViewController provides
support for this.

this initial implementation was tested on the simulator. CPU usage
during resizing still needs to be measured on a real device. also drag
handles are only drawn in a rudimentary fashion at the moment.