remove GoMove reference to "next"
with a tree of nodes model, a move can have many successor (next) moves,
each of which is located in a separate branch (game variation) of the
tree. an unambigous successor (next) move is only possible within the
same game variation. keeping the "next" property in GoMove would require
its value be changed whenever the user changes the current game
variation - this would be much too prone to errors, so removing the
"next" property entirely is the cleanest solution.

change details:
- the complicated maintenance of the next/previous references in
  GoMove's convenience constructor and the dealloc methhod is no longer
  necessary (and in fact no longer possible).
- a number of unit tests that test this behaviour can therefore be
  removed (including the tests for the recent bugfixes of GitHub issues
  #369 and #370).
- the GoUtilities method relinkMoves(), which relinks moves after
  unarchiving, is much simpler now; the method was also rewritten to
  relink the entire tree of nodes, i.e. full game variation support.
- the GoUtilities method nodeWithNextMove() was rewritten to operate on
  the current game variation instead of operating on the direct
  descendants of a node. the method was renamed to clearly indicate how
  it works.
- a number of call sites of the renamed method needed to be changed
  accordingly.
- the few remaining actual usages of the GoMove property "next" were
  rewritten to determine the next move via the renamed GoUtilities
  method