resizable panes in ResizableStackViewController can no longer grow larger than 100%
before this change, the effective size on screen of all resizable panes
could grow to be larger than the size of the ResizableStackViewController
main view: if one or more subviews in the pane to shrink has a size
constraint that doesn't let that subview shrink (e.g. a fixed size),
Auto Layout keeps the effective size of the resizable pane at the
minimum allowed size. Because ResizableStackViewController is not aware
of externally applied minimum size constraint, it thinks it can shrink
the pane that is already at the minimum size, and grow the other pane.

while Auto Layout prevents the shrinking, it happily follows the
instructions of ResizableStackViewController to grow the other pane.
up until now this resulted in an effective total size of more than 100%.

the change simply consists of connecting the last resizable pane to the
superview edge, and to give this connecting constraint a higher priority
than the sizing constraints. Auto Layout now knows that it must break the
sizing constraint if it would violate the 100% rule.