fix layout on Archive tab
the table view on the Archive tab was incorrectly laid out, with the
content being visible when scrolled behind the navigation bar at the top
or the tab bar at the bottom. the simplest approach to get it right in
all iOS versions and on all devices is to let UITableViewController
handle the stuff.

ArchiveViewController therefore now uses UITableViewController instead
of UITableView. after this change ArchiveViewController is now a
container view controller.