LoadGameCommand now properly informs the rest of the application about the final Go model state
this fixes at least one issue: when the user loads a game the board
view drawing layers are notified about creation of the new game, but
then they were not notified about the final number of board positions
and the final current board position. when the user then switched to the
first board position (rewind) some of the layers would not redraw
because they could not detect a change in the board state (the last
notification they had was for the newly created game which did not have
any board positions).