use default tab bar appearance on all devices
the rationale for showing a black tab bar on small iPhones was that the
default translucent appearance of the tab bar does not look good when
a toolbar is stacked on top of the tab bar.

because nowadays there is no toolbar stacked on top of the tab bar
anymore, the rationale for a black tab bar has disappeared.