fix NSSecureCoding implementation of unarchiving (#409)
in commit 5345ff8ca82c79dd18c024aa0cadfbf32de848e3 the NSCoding
implementation was adapted to NSSecureCoding. a few things were
overlooked so that the commit did not result in a working unarchiving
solution. this commit contains the fixes:

- the root object GoGame is now decoded with the new
  decodeObjectOfClass:forKey:() instead of with the old
  decodeObjectForKey() - in essence the same approach is used as in
  the various initWithCoder:() implementations
- when initWithCoder:() implementations decode a collection object
  (e.g. NSArray, NSDictionary) they now supply not only the collection
  class but also the class(es) of the collection items
  => decodeObjectOfClasses:forKey() is now used instead of
     decodeObjectOfClass:forKey()
- the initWithCoder:() implementation in GoBoard still used
  decodeObjectForKey:() - this is now fixed

a change rather than a bugfix is that unarchiving now uses
NSDecodingFailurePolicyRaiseException instead of the default
NSDecodingFailurePolicySetErrorAndReturn, in order to improve diagnostics
if unarchiving fails.