eliminate use of GoGame class method sharedGame() in the Go module
using the GoGame class method sharedGame() adds a dependeny on global
state. this must be prevented within the classes that live in the Go
module because they must be self-contained - basically we want to be
sure that at runtime Go objects in a GoGame object tree can only refer
to other Go objects that live in the same object tree.

if a class does not have its own reference that somehow points into a
GoGame object tree, the class is not allowed to obtain the reference by
itself by invoking GoGame sharedGame(). instead an outside source must
provide the needed reference via a parameter.