NodeTreeViewCanvasTest: create testee only after node tree is set up
this change is intended to make sure that canvas calculation in the
testee is not prematurely triggered by some notification posted by one
of the Go objects during node tree setup.

this should not be the case in the tests modified by this commit, but
in tests added after this commit the order will be important. this
commit merely aligns the layout of existinng tests with the layout of
upcoming new tests.