StatusViewController is no longer a UIViewController => fixes a layout error on iPad (#304)
the trigger for this change is that on iPad devices on iOS 26, when the
window size is changed so that layout orientation switches from
landscape to portrait, the status view is not properly laid out (it
covers most of the Go board and lets it appear blurry through some sort
of half-transparency).

the root cause is that in Portrait layout the status view is assigned as
title view of the navigation item. this then causes some sort of clash
between the UINavigationController that tries to render the status view
as part of the navigation view, and StatusViewController which until now
was a UIViewController.

when StatusViewController is no longer a UIViewController the clash
disappears and UINavigationController gets full control of embedding the
status view.

fun fact: once upon a time StatusViewController already was just a
simple NSObject. this was changed and StatusViewController became a
UIViewController in commit 6fe75e8f662635268096e4e7f2e8f2b1db617713
(in 2013).