change implementation of two methods in PathUtilities
the changes were originally intended to reduce the number of events
received by FilesystemMonitor. after the changes were implemented it
turned out that the events triggered by the NSFileManager method
replaceItemAtURL:withItemAtURL:backupItemName:options:resultingItemURL:error:()
still were not ideal.

as a new alternative, FilesystemOperations was invented (see previous
commit), but the changes in PathUtilities are still deemed to be useful,
so they are made permanent by this commit.