introduce notification currentBoardPositionWillChanged
introduce notification currentBoardPositionWillChanged, to complement
the existing currentBoardPositionDidChange; the new notification was
originally thought to be necessary for the time handling when the
selected node changes, but it turned out not to be. because the
notification might prove useful in the future, and introducing it was
substantial work, we keep the notification for the moment.